+++ CHANGELOG.md
@@ -8,6 +8,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 * Made LogEntry implement Serializable.
 
+### Changed
+* Send the correct Content-Type header for API versioning.
+
 ## [0.6.3-beta](https://github.com/contentful/contentful.php/tree/0.6.3-beta) (2016-03-02)
 
 ### Added